split classes and fix bug in request get passenger by email
