+++ src/main/java/com/indytskyi/userserviceairport/controller/AuthenticationController.java
@@ -2,6 +2,7 @@ package com.indytskyi.userserviceairport.controller;
 
 import com.indytskyi.userserviceairport.dto.*;
 import com.indytskyi.userserviceairport.service.AuthenticationService;
+import com.indytskyi.userserviceairport.service.RegistrationService;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -14,10 +15,11 @@ import org.springframework.web.bind.annotation.*;
 @RequiredArgsConstructor
 public class AuthenticationController {
   private final AuthenticationService authenticationService;
+  private final RegistrationService registrationService;
 
   @PostMapping("/register")
   public ResponseEntity<RegisterResponseDto> register(@RequestBody @Valid RegisterRequest request) {
-    return new ResponseEntity<>(authenticationService.register(request),
+    return new ResponseEntity<>(registrationService.register(request),
             HttpStatus.CREATED);
   }
   @PostMapping("/login")
@@ -28,12 +30,12 @@ public class AuthenticationController {
 
   @GetMapping("/confirm")
   public ResponseEntity<String> confirm(@RequestParam("token") String token) {
-    return ResponseEntity.ok(authenticationService.confirmToken(token));
+    return ResponseEntity.ok(registrationService.confirmToken(token));
   }
 
   @GetMapping("/resend-confirm-email")
   public ResponseEntity<Object> resendConfirmEmail(@RequestParam String email) {
-    return ResponseEntity.ok(authenticationService.resendEmail(email));
+    return ResponseEntity.ok(registrationService.resendEmail(email));
   }
 
   @GetMapping("/refresh-token")
