+++ src/main/java/com/indytskyi/userserviceairport/dto/ValidateTokenResponseDto.java
@@ -1,0 +1,4 @@
+package com.indytskyi.userserviceairport.dto;
+
+public record ValidateTokenResponseDto(Long userId, String role) {
+}
