+++ src/main/java/com/indytskyi/userserviceairport/service/AuthenticationService.java
@@ -4,11 +4,6 @@ import com.indytskyi.userserviceairport.dto.*;
 import com.indytskyi.userserviceairport.model.User;
 
 public interface AuthenticationService {
-    RegisterResponseDto register(RegisterRequest request);
     AuthenticationResponse authenticate(AuthenticationRequest request);
-    String confirmToken(String token);
-
-    Object resendEmail(String email);
-
     AuthenticationResponse refreshToken(RefreshTokenRequestDto refreshTokenRequestDto);
 }
